move custom search dialog to its own component separate from default search behavior
