+++ app/layout.config.tsx
@@ -3,6 +3,7 @@ import { Github } from "lucide-react";
 import { CommunityButton } from "../components/Community";
 import { NavButton } from "../components/NavButton";
 import { XButton } from "../components/XButton";
+import RAGSearchToggle from "../components/RAGSearchToggle";
 
 /**
  * Shared layout configurations
@@ -43,6 +44,7 @@ export const baseOptions: BaseLayoutProps = {
 			on: "nav",
 			children: (
 				<div className="flex items-center gap-2 text-sm">
+					<RAGSearchToggle />
 					<NavButton href="https://agentuity.com/blog" target="_blank" rel="noopener noreferrer">Blog</NavButton>
 					<NavButton href="https://app.agentuity.com" target="_blank" rel="noopener noreferrer">Console</NavButton>
 					<CommunityButton target="_blank" rel="noopener noreferrer" />
